requirements.txt
asgiref==3.8.1
Django==4.2.20
django-discover-runner==1.0
django-role-permissions==3.2.0
pillow==11.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
